Rename to `rustServiceTypes`
