+++ src/services/RustService/lib/fetchDataBuildInfo.ts
@@ -1,5 +1,5 @@
 import callRustService from "../callRustService";
-import type { RustServiceDataBuildInfo } from "../types";
+import type { RustServiceDataBuildInfo } from "../rustServiceTypes";
 
 export default async function fetchDataBuildInfo(): Promise<RustServiceDataBuildInfo> {
   return callRustService<RustServiceDataBuildInfo>("get_data_build_info");
