+++ src/services/RustService/lib/fetchImageInfo.ts
@@ -1,5 +1,5 @@
 import callRustService from "../callRustService";
-import type { RustServiceImageInfo } from "../types";
+import type { RustServiceImageInfo } from "../rustServiceTypes";
 
 export default async function fetchImageInfo(
   filename: string,
